+++ Package.resolved
@@ -42,8 +42,8 @@
         "repositoryURL": "https://github.com/OperatorFoundation/Song.git",
         "state": {
           "branch": null,
-          "revision": "f004b968df1246a17f1ee94ac8a9b88beda871c7",
-          "version": "0.0.17"
+          "revision": "0da25c70c7954274ffd4aec08fec46dd37045b19",
+          "version": "0.0.18"
         }
       },
       {
