+++ Package.swift
@@ -16,8 +16,7 @@ let package = Package(
         .package(url: "https://github.com/OperatorFoundation/SwiftQueue.git", from: "0.0.3"),
         .package(url: "https://github.com/OperatorFoundation/SwiftPCAP.git", from: "1.1.7"),
         .package(url: "https://github.com/OperatorFoundation/Datable.git", from: "3.0.2"),
-        .package(name: "Rethink", url: "https://github.com/OperatorFoundation/rethink-swift.git", from: "1.0.0"),
-        .package(url: "https://github.com/OperatorFoundation/Song.git", from: "0.0.17"),
+        .package(url: "https://github.com/OperatorFoundation/Song.git", from: "0.0.19"),
         .package(url: "https://github.com/apple/swift-argument-parser.git", from: "0.0.1"),
         .package(url: "https://github.com/OperatorFoundation/Bits.git", from: "1.0.3"),
         .package(url: "https://github.com/OperatorFoundation/InternetProtocols.git", from: "1.0.1"),
